Add files via upload
To use this code, you will need a Genius.com account, which you will use to create an API Token. Instructions on how to do that can be found here: https://docs.genius.com/

This will program create a json which you will run and it will create a txt file of lyrics.